finish major redux file struct refactor
