--- eventkit_cloud/ui/static/ui/app/actions/actionTypes.js
@@ -1,38 +1,0 @@
-const types = {
-    LOAD_RUNS_SUCCESS: 'LOAD_JOBS_SUCCESS',
-    DRAW_SEARCH_BBOX: 'DRAW_SEARCH_BBOX',
-    CLEAR_SEARCH_BBOX: 'CLEAR_SEARCH_BBOX',
-    FETCHING_RUNS: 'FETCHING_RUNS',
-    RECEIVED_RUNS: 'RECEIVED_RUNS',
-    FETCH_RUNS_ERROR: 'FETCH_RUNS_ERROR',
-    FETCHING_FEATURED_RUNS: 'FETCHING_FEATURED_RUNS',
-    RECEIVED_FEATURED_RUNS: 'RECEIVED_FEATURED_RUNS',
-    FETCH_FEATURED_RUNS_ERROR: 'FETCH_FEATURED_RUNS_ERROR',
-    LOAD_JOBS_SUCCESS: 'LOAD_JOBS_SUCCESS',
-    DELETING_RUN: 'DELETING_RUN',
-    DELETED_RUN: 'DELETED_RUN',
-    DELETE_RUN_ERROR: 'DELETE_RUN_ERROR',
-    GETTING_DATACART_DETAILS: 'GETTING_DATACART_DETAILS',
-    CLEAR_DATACART_DETAILS: 'CLEAR_DATACART_DETAILS',
-    DATACART_DETAILS_RECEIVED: 'DATACART_DETAILS_RECEIVED',
-    DATACART_DETAILS_RECEIVED_FLAG: 'DATACART_DETAILS_RECEIVED_FLAG',
-    DATACART_DETAILS_ERROR: 'DATACART_DETAILS_ERROR',
-    RERUNNING_EXPORT: 'RERUNNING_EXPORT',
-    RERUN_EXPORT_ERROR: 'RERUN_EXPORT_ERROR',
-    RERUN_EXPORT_SUCCESS: 'RERUN_EXPORT_SUCCESS',
-    CLEAR_RERUN_INFO: 'CLEAR_RERUN_INFO',
-    CANCELING_PROVIDER_TASK: 'CANCELING_PROVIDER_TASK',
-    CANCELED_PROVIDER_TASK: 'CANCELED_PROVIDER_TASK',
-    CANCEL_PROVIDER_TASK_ERROR: 'CANCEL_PROVIDER_TASK_ERROR',
-    UPDATING_EXPIRATION: 'UPDATING_EXPIRATION',
-    UPDATE_EXPIRATION_ERROR: 'UPDATE_EXPIRATION_ERROR',
-    UPDATE_EXPIRATION_SUCCESS: 'UPDATE_EXPIRATION_SUCCESS',
-    UPDATING_PERMISSION: 'UPDATING_PERMISSION',
-    UPDATE_PERMISSION_ERROR: 'UPDATE_PERMISSION_ERROR',
-    UPDATE_PERMISSION_SUCCESS: 'UPDATE_PERMISSION_SUCCESS',
-    SET_PAGE_ORDER: 'SET_PAGE_ORDER',
-    SET_PAGE_VIEW: 'SET_PAGE_VIEW',
-};
-
-export default types;
-
