+++ eventkit_cloud/ui/static/ui/app/reducers/dataPackReducer.js
@@ -1,5 +1,79 @@
-import types from '../actions/actionTypes';
-import initialState from './initialState';
+import { types } from '../actions/datapackActions';
+import { types as uiTypes } from '../actions/uiActions';
+
+export const initialState = {
+    runsList: {
+        fetching: false,
+        fetched: false,
+        runs: [],
+        error: null,
+        nextPage: false,
+        range: '',
+        order: '',
+        view: '',
+        cancelSource: null,
+    },
+    featuredRunsList: {
+        fetching: false,
+        fetched: false,
+        runs: [],
+        error: null,
+        nextPage: false,
+        range: '',
+        cancelSource: null,
+    },
+    runDeletion: {
+        deleting: false,
+        deleted: false,
+        error: null,
+    },
+    updateExpiration: {
+        updating: false,
+        updated: false,
+        error: null,
+    },
+    datacartDetails: {
+        fetching: false,
+        fetched: false,
+        data: [],
+        error: null,
+    },
+};
+
+export function getDatacartDetailsReducer(state = initialState.datacartDetails, action) {
+    switch (action.type) {
+        case types.GETTING_DATACART_DETAILS:
+            return {
+                ...state,
+                fetching: true,
+                fetched: false,
+                error: null,
+            };
+        case types.DATACART_DETAILS_RECEIVED:
+            return {
+                fetching: false,
+                fetched: true,
+                data: action.datacartDetails.data,
+                error: null,
+            };
+        case types.DATACART_DETAILS_ERROR:
+            return {
+                fetching: false,
+                fetched: false,
+                data: [],
+                error: action.error,
+            };
+        case types.CLEAR_DATACART_DETAILS:
+            return {
+                fetching: false,
+                fetched: false,
+                data: [],
+                error: null,
+            };
+        default:
+            return state;
+    }
+}
 
 export function dataPackReducer(state = initialState.runsList, action) {
     switch (action.type) {
@@ -22,9 +96,9 @@ export function dataPackReducer(state = initialState.runsList, action) {
             return {
                 ...state, fetching: false, fetched: false, runs: [], error: action.error, cancelSource: null,
             };
-        case types.SET_PAGE_ORDER:
+        case uiTypes.SET_PAGE_ORDER:
             return { ...state, order: action.order };
-        case types.SET_PAGE_VIEW:
+        case uiTypes.SET_PAGE_VIEW:
             return { ...state, view: action.view };
         default:
             return state;
@@ -57,7 +131,7 @@ export function featuredRunsReducer(state = initialState.featuredRunsList, actio
     }
 }
 
-export function DeleteRunsReducer(state = initialState.runsDeletion, action) {
+export function deleteRunReducer(state = initialState.runDeletion, action) {
     switch (action.type) {
         case types.DELETING_RUN:
             return { deleting: true, deleted: false, error: null };
@@ -65,6 +139,19 @@ export function DeleteRunsReducer(state = initialState.runsDeletion, action) {
             return { deleting: false, deleted: true, error: null };
         case types.DELETE_RUN_ERROR:
             return { deleting: false, deleted: false, error: action.error };
+        default:
+            return state;
+    }
+}
+
+export function updateExpirationReducer(state = initialState.updateExpiration, action) {
+    switch (action.type) {
+        case types.UPDATING_EXPIRATION:
+            return { updating: true, updated: false, error: null };
+        case types.UPDATE_EXPIRATION_SUCCESS:
+            return { updating: false, updated: true, error: null };
+        case types.UPDATE_EXPIRATION_ERROR:
+            return { updating: false, updated: false, error: action.error };
         default:
             return state;
     }
