--- eventkit_cloud/ui/static/ui/app/reducers/initialState.js
@@ -1,85 +1,0 @@
-
-export default {
-    aoiInfo: {
-        geojson: {},
-        originalGeojson: {},
-        geomType: null,
-        title: null,
-        description: null,
-        selectionType: null,
-        buffer: 0,
-    },
-    drawer: 'closed',
-    runsList: {
-        fetching: false,
-        fetched: false,
-        runs: [],
-        error: null,
-        nextPage: false,
-        range: '',
-        order: '',
-        view: '',
-        cancelSource: null,
-    },
-    featuredRunsList: {
-        fetching: false,
-        fetched: false,
-        runs: [],
-        error: null,
-        nextPage: false,
-        range: '',
-        cancelSource: null,
-    },
-    runsDeletion: {
-        deleting: false,
-        deleted: false,
-        error: null,
-    },
-    submitJob: {
-        fetching: false,
-        fetched: false,
-        jobuid: '',
-        error: null,
-    },
-    exportInfo: {
-        exportName: '',
-        datapackDescription: '',
-        projectName: '',
-        providers: [],
-        areaStr: '',
-        formats: ['gpkg'],
-    },
-    stepperNextEnabled: false,
-    datacartDetails: {
-        fetching: false,
-        fetched: false,
-        data: [],
-        error: null,
-    },
-    runDeletion: {
-        deleting: false,
-        deleted: false,
-        error: null,
-    },
-    exportReRun: {
-        fetching: false,
-        fetched: false,
-        data: [],
-        error: null,
-    },
-    cancelProviderTask: {
-        canceling: false,
-        canceled: false,
-        error: null,
-    },
-    updateExpiration: {
-        updating: false,
-        updated: false,
-        error: null,
-    },
-    updatePermission: {
-        updating: false,
-        updated: false,
-        error: null,
-    },
-};
