+++ eventkit_cloud/ui/static/ui/app/reducers/licenseReducer.js
@@ -1,6 +1,6 @@
 import { types } from '../actions/licenseActions';
 
-const initialState = {
+export const initialState = {
     fetching: false,
     fetched: false,
     licenses: [],
